--- src/main/java/blackjack/domain/PlayerProfitResultDto.java
@@ -1,8 +1,0 @@
-package blackjack.domain;
-
-public record PlayerProfitResultDto(String playerName, double profit) {
-
-    public static PlayerProfitResultDto from(Player player, Double profit) {
-        return new PlayerProfitResultDto(player.getName(), profit);
-    }
-}
