+++ src/main/java/blackjack/dto/PlayerProfitResult.java
@@ -1,0 +1,10 @@
+package blackjack.dto;
+
+import blackjack.domain.Player;
+
+public record PlayerProfitResult(String playerName, double profit) {
+
+    public static PlayerProfitResult from(Player player, Double profit) {
+        return new PlayerProfitResult(player.getName(), profit);
+    }
+}
